Fix broken, crash-causing French translation.
The translation of IntroductionText had two occurrences of %@ in it, causing
stringWithFormat: to crash in askCrashReportDetails.  Needless to say, crashing
on application startup when trying to report a crash and preventing the
application from ever successfully launching again is less than ideal.

This could be fixed by using %1$@ for the second occurrence, but repeating the
name multiple times is not ideal, so this commit replaces it with "to the
application's developers" instead, as suggested by a native French speaker.