Fix finding latest feedback message having an ID
The previous implementation stopped at the first message having an ID instead of, as the method name suggests, getting the last item.

This change fixes the detection of knowing if there are new messages incoming from the server the user didn’t see yet and hence makes the use of the `showAlertOnIncomingMessages` actually possible.